AI learning flappy bird/requirements.txt
pygame==2.5.2
numpy==1.24.3
neat-python==0.92 